ARC: int vs long
 - arch specific bitsperlong.h
 - general LP64 related changes
 - don't force cast argument to __builtin_arc_sr()
 - irqflags: STAT32 is 32-bits but generic raw_local_irq_* expect flags
             to be long

+++ arch/arc/include/asm/mmu.h
@@ -79,7 +79,7 @@
 #ifndef __ASSEMBLY__
 
 typedef struct {
-	unsigned long asid[NR_CPUS];	/* 8 bit MMU PID + Generation cycle */
+	unsigned int asid[NR_CPUS];	/* 8 bit MMU PID + Generation cycle */
 } mm_context_t;
 
 #ifdef CONFIG_ARC_DBG_TLB_PARANOIA
